+++ src/core/solvers/tier_solver/reverse_tier_graph.c
@@ -7,8 +7,8 @@
  * @author GamesCrafters Research Group, UC Berkeley
  *         Supervised by Dan Garcia <ddgarcia@cs.berkeley.edu>
  * @brief Implementation of the ReverseTierGraph type.
- * @version 1.1.0
- * @date 2024-03-27
+ * @version 1.1.1
+ * @date 2024-12-22
  *
  * @copyright This file is part of GAMESMAN, The Finite, Two-person
  * Perfect-Information Game Generator released under the GPL:
@@ -78,7 +78,7 @@ TierArray ReverseTierGraphPopParentsOf(ReverseTierGraph *graph, Tier child) {
     TierHashMapIterator it = TierHashMapGet(&graph->index_of, child);
     if (!TierHashMapIteratorIsValid(&it)) return kIllegalTierArray;
 
-    int index = TierHashMapIteratorValue(&it);
+    int64_t index = TierHashMapIteratorValue(&it);
     TierArray ret = graph->parents_of[index];
     memset(&graph->parents_of[index], 0, sizeof(graph->parents_of[index]));
 
@@ -94,7 +94,7 @@ TierArray ReverseTierGraphGetParentsOf(ReverseTierGraph *graph, Tier child) {
     TierHashMapIterator it = TierHashMapGet(&graph->index_of, child);
     if (!TierHashMapIteratorIsValid(&it)) return kIllegalTierArray;
 
-    int index = TierHashMapIteratorValue(&it);
+    int64_t index = TierHashMapIteratorValue(&it);
     TierArray ret;
     TierArrayInitCopy(&ret, &graph->parents_of[index]);
